Fix a good portion of the routing logic (#28)
* Initializing module to handle cost

* cfg: Adding dependency on serde + serde_json

* cfg: Loading module cost

* style:

* refact + style: Renaming to CostFunction

* style:

* style:

* refact: Lowering multiplier_scalar to f32

For now it makes everything easier since the features are expected to be
f32 as well.

* test, fix: Renaming features names

* feat: cost::sample::cost_function()

* refact: CostFunction should be pub(crate)

* refact: cost::sample should be pub(crate)

* feat: cost::CostFunction::calculate_chunk()

Calculate cost in a chunk.

* style:

* fix: Missing imports

* Adding cost_function as component of Dataset

* style:

* Re-defining log level

* cost_function as an item of Dataset

* refact: Using cost calculation from CostFunction

* style:

* fix: Adding cost_function as an item of Dataset

* fix: Missing import debug log level

* Confirming with cost_function as a new item of Dataset

* log: Improving log message

* fix: Avoiding hardcoded varname for chunk shape estiamte

For now we are repeating the input chunk shape, and assume all variables
have the same shape. The varname was hardcoded, but now it takes the
first one in the list (a temporary solution).

* log: Adjusting log levels

* style:

* cfg: Updating Cargo.lock

* cfg: Bump patch version

* fix: into_path() is deprecated

* Fix broken tag

* Rename python modules

* Add minor test detail

* Update lockfile

* Add identifier

* Add new dev deps

* Implement from train for tuple to Point

* Update docstring

* New Ruff ignore

* Add pyo3 dep

* Rename rust library

* Update lib name

* MVP of working maturin

* Don't block lib folder

* MVP of FFI module

* Remove unused config option

* Use `_rust` for module name

* Revert back to old name

* Update maturin version

* Add version to init file

* Update pixi lock

* Fix import

* Add zarr req

* Add most basic rust bindings test

* impl From<Point> for (u64, u64)

* Add in computation

* Make use of default param

* Allow multiple starting points

* Fix call

* Add function signature and docstring

* Rename function

* Change name

* Update docstring

* Update name

* Fix typo

* Let cost be mutable so it's not dropped after the let clause

* Use multiplier layer name

* get_3x3 takes point

* Fix typo

* Fix typo

* FIx x and y backwards on point neighbors

* Add tests

* Don't scale cost

* Add skimage for ground truth tests

* Add `PRECISION_SCALAR`

* Cost now returned as float

* Add cost scale and unscale functions

* Add todo item

* Update to test against MCP Geometric

* Remove spurious prints

* Remove completed todo item

* Linter fix

* Trying to fix macos test

* More attempts

* Update

* Update

* Trying to fix mac install errors

* Trying more things

* Try more things

* Try more

* try more

* Try lint fix

* Remove comments

* Update pyo3 features

* Update lock file

* Update lock file

* Add back gitignore

* Update package path

* Remove from impl

* Rename struct

* Add test

* (WIP) rename python module

* Rename folder

* use correct package name

* Update lock file

* Remove not necessary ruff rule

* Linter fixes

* Testing

* Fix struct name

* Update lock file

* Use u64 for cost tracking

* Add todo

* Add `new` func

---------

+++ crates/revrt/src/ffi.rs
@@ -79,7 +79,7 @@ fn find_paths(
     start: Vec<(u64, u64)>,
     end: Vec<(u64, u64)>,
     cache_size: u64,
-) -> Result<Vec<(Vec<(u64, u64)>, usize)>> {
+) -> Result<Vec<(Vec<(u64, u64)>, f32)>> {
     let start: Vec<ArrayIndex> = start
         .into_iter()
         .map(|(i, j)| ArrayIndex { i, j })
